project-root/backend/requirements.txt
Django>=4.2
djangorestframework
Pillow
numpy
# Optional but recommended for face recognition
face_recognition
opencv-python